专项检查napi中的无业务逻辑的nullptr Signed-off-by: zhaoqil <zhaoqi70@huawei.com>

+++ service/filter/src/event_filter_proxy.cpp
@@ -54,7 +54,7 @@ bool EventFilterProxy::HandlePointerEvent(const std::shared_ptr<PointerEvent> ev
         return false;
     }
 
-    const uint32_t code = static_cast<uint32_t>(OPERATOR_TYPE::HANDLE_POINTER_EVENT);
+    const uint32_t code = OPERATOR_TYPE::HANDLE_POINTER_EVENT;
     int32_t requestResult = Remote()->SendRequest(code, data, reply, option);
     if (requestResult != NO_ERROR) {
         MMI_LOGE("send request fail, result:%{public}d", requestResult);
